refactored training scripts, emphasis on translation examples
