--- rcl_dataset_tatoeba/English-Yoruba.tsv
@@ -1,45 +1,0 @@
-﻿1860	I don't want to go to school.	3633484	Mi ò fẹ́ lọ sí ilé-ẹ́kọ́.
-256018	I'm a doctor.	3143489	Onísègùn ni mí.
-1977717	I want to go.	3335168	Mo fẹ́ lọ.
-2412739	I have two cats.	3081443	Mo ní ológbò méjì.
-5257037	The horse is white.	3143475	Ẹṣin jẹ́ funfun.
-3143476	The horse is not white.	3143477	Ẹṣin kìí ṣe funfun.
-3143478	This horse is white.	3143483	Ẹṣin yìí jẹ́ funfun.
-3143479	That horse is white.	3143485	Ẹṣin náà jẹ́ funfun.
-3143480	This horse is not white.	3143486	Ẹṣin yìí kìí ṣe funfun.
-3143482	That horse is not white.	3143487	Ẹṣin náà kìí ṣe funfun.
-3143490	You're a doctor.	3143491	Onísègùn ni ẹ́.
-5226593	He's a doctor.	3143493	Onísègùn ni.
-3143495	She's a doctor.	3143493	Onísègùn ni.
-3143497	We're doctors.	3143498	Onísègùn ni wá.
-3143499	You're doctors.	3143500	Onísègùn ni yín.
-3143502	They're doctors.	3143503	Onísègùn ni wọ́n.
-3154933	My wife went to the farm.	3154934	Ìyàwó mi lọ sí oko.
-3154937	His money is in the bank.	3154939	Owó rẹ̀ wà ní bánkì.
-3154996	It surprises me.	3154998	Ó yà mí l'ẹ́nu.
-3179083	It's midday.	3179086	Agogo mejila ọsan.
-3180912	I want to buy a book.	3180913	Mo fẹ́ rà ìwé.
-3188277	The book isn't new.	3197736	Ìwé kìí ṣe titun.
-373320	How are you?	3348007	Bawo ni?
-3352026	One shouldn't injure during the day, the body side on which one will lie at night.	3352025	Ẹ̀gbẹ́ táa máa fi sùn lálẹ́, kò yẹ ká fi gbọgbẹ́ lọ́ọ̀sán.
-3352033	It's one who shares in another's wealth that hails him as a generous person.	3352030	Ẹni tó bá ńjẹ nínú ọlà ẹnìkan nií pèé ní anímáṣaun.
-2064575	Who is Emily?	3713961	Ta ni Emily?
-3663249	There are islands in the sea.	3713991	Àwọn erékùṣù wà ní òkun.
-773503	I have brothers.	3714005	Mo ní àwọn arákunrin.
-1434	I love you.	3808207	Mo nifẹẹ rẹ.
-1399479	Burj Khalifa is currently the tallest skyscraper in the world.	4266589	Burj Khalifa je agbar tó ga ju ni gbògbó ayè.
-1646	My name is Jack.	4535812	Jack lorúkọ mi.
-4536791	Did you go to the market yesterday?	4536795	Ṣé o lọ sọjà ni àna?
-477086	Yes.	4536797	Bẹ́ẹ̀ni.
-4536799	I went to the river.	4536800	Mo lọ sí odò.
-373216	Really?	4641859	Gidi gan?
-3861611	Water is life.	5978173	Omi ni ayé.
-611372	We want complete sentences.	8773821	A fẹ́ àwọn ọ̀rọ̀ ní gbólóhùn pípé.
-2700686	There is no problem.	10726770	kò sí wahála.
-6110256	Tom didn't break anything.	10739589	Tom ko fó nkan kan
-10739161	Welcome back home.	10739593	Kaabo padà sí ilé
-2258234	Good morning.	10963734	Ẹ káǎrọ̀.
-2258234	Good morning.	10963735	Káǎrọ̀.
-1849392	Be careful.	10963805	Ẹ rọra.
-1646	My name is Jack.	10963865	Orúkọ mi ni Jack.
-448645	Thanks!	10963871	Ẹ ṣé!
